pyproject.toml
[project]
name = "credit-card-scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.13.4",
    "pymupdf==1.25.5",
    "fuzzywuzzy==0.18.0",
    "groq==0.23.1",
    "openai==1.76.0",
    "pymupdf>=1.25.5",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "streamlit==1.44.1",
]
